создал класс Api для общения с сервером и использовал его для первичной загрузки изображений
